Clean up dc events and add month_range to utils
